Add import text and moved to Assignment test file
