+++ tests/Feature/ImportExport/Exporters/ProcessExporterTest.php
@@ -157,101 +157,4 @@ class ProcessExporterTest extends TestCase
         $this->assertEquals($subProcess->id, Utils::getPmConfig($element)['processId']);
         $this->assertEquals(0, Process::where('name', 'package')->count());
     }
-
-    /**
-     * @group fix
-     */
-    public function testExportImportAssignments()
-    {
-        // Create users and groups
-        $users = User::factory(12)->create();
-        $groups = Group::factory(10)->create();
-
-        // Assign three users to group 1, assign two users to group 2, assign one user to group 3
-        foreach ($users as $key => $user) {
-            if ($key <= 2) {
-                $group = $groups[0];
-            }
-            if ($key > 2 and $key <= 4) {
-                $group = $groups[1];
-            }
-            if ($key > 4 and $key <= 5) {
-                $group = $groups[2];
-            }
-
-            // Assign last user to last group
-            if ($key == 11) {
-                $group = $groups[9];
-            }
-
-            if ($key > 5) {
-                continue;
-            }
-
-            GroupMember::factory()->create([
-                'member_type' => User::class,
-                'member_id' => $user->id,
-                'group_id' => $group->id,
-            ]);
-        }
-
-        $this->addGlobalSignalProcess();
-
-        // Create process
-        $process = $this->createProcess('process-with-different-kinds-of-assignments', ['name' => 'processTest']);
-
-        // Assign users to process assignments
-        Utils::setAttributeAtXPath($process, '/bpmn:definitions/bpmn:process/bpmn:task[1]', 'pm:assignedUsers', implode(',', [$users[0]->id, $users[1]->id, $users[2]->id]));
-        Utils::setAttributeAtXPath($process, '/bpmn:definitions/bpmn:process/bpmn:task[2]', 'pm:assignedUsers', implode(',', [$users[3]->id, $users[4]->id]));
-        Utils::setAttributeAtXPath($process, '/bpmn:definitions/bpmn:process/bpmn:manualTask[1]', 'pm:assignedUsers', implode(',', [$users[5]->id, $users[6]->id]));
-        Utils::setAttributeAtXPath($process, '/bpmn:definitions/bpmn:process/bpmn:manualTask[2]', 'pm:assignedUsers', implode(',', [$users[7]->id]));
-        Utils::setAttributeAtXPath($process, '/bpmn:definitions/bpmn:process/bpmn:callActivity[1]', 'pm:assignedUsers', implode(',', [$users[8]->id, $users[9]->id]));
-        Utils::setAttributeAtXPath($process, '/bpmn:definitions/bpmn:process/bpmn:callActivity[2]', 'pm:assignedUsers', implode(',', [$users[10]->id]));
-
-        // Assign groups to process assignments
-        Utils::setAttributeAtXPath($process, '/bpmn:definitions/bpmn:process/bpmn:task[1]', 'pm:assignedGroups', implode(',', [$groups[0]->id]));
-        Utils::setAttributeAtXPath($process, '/bpmn:definitions/bpmn:process/bpmn:task[2]', 'pm:assignedGroups', implode(',', [$groups[1]->id, $groups[2]->id]));
-        Utils::setAttributeAtXPath($process, '/bpmn:definitions/bpmn:process/bpmn:manualTask[1]', 'pm:assignedGroups', implode(',', [$groups[3]->id]));
-        Utils::setAttributeAtXPath($process, '/bpmn:definitions/bpmn:process/bpmn:manualTask[2]', 'pm:assignedGroups', implode(',', [$groups[4]->id, $groups[5]->id]));
-        Utils::setAttributeAtXPath($process, '/bpmn:definitions/bpmn:process/bpmn:callActivity[1]', 'pm:assignedGroups', implode(',', [$groups[6]->id, $groups[7]->id, $groups[8]->id]));
-        Utils::setAttributeAtXPath($process, '/bpmn:definitions/bpmn:process/bpmn:callActivity[2]', 'pm:assignedGroups', implode(',', [$groups[9]->id]));
-
-        $process->save();
-
-        $this->runExportAndImport($process, ProcessExporter::class, function () use ($process) {
-            User::query()->forceDelete();
-            Group::query()->forceDelete();
-            GroupMember::query()->forceDelete();
-            Process::query()->forceDelete();
-
-            $this->assertEquals(0, User::get()->count());
-            $this->assertEquals(0, Group::get()->count());
-            $this->assertEquals(0, GroupMember::get()->count());
-            $this->assertEquals(0, Process::get()->count());
-        });
-
-        // 11 from 12 created users should be exported ..
-        $this->assertEquals(11, User::whereIn('username', $users->pluck('username'))->get()->count());
-        $this->assertEquals(10, Group::whereIn('name', $groups->pluck('name'))->get()->count());
-        $this->assertDatabaseHas('processes', ['name' => $process->name]);
-        $process = Process::where('name', $process->name)->firstOrFail();
-
-        // reasignar los ids de los usuarios nuevos al bpmn del nuevo proceso
-        dd($process->bpmn);
-        // fijarse en $process->bpmn y hacer assert que existe
-        // Utils::setAttributeAtXPath($process, '/bpmn:definitions/bpmn:process/bpmn:task[1]', 'pm:assignedUsers', implode(',', [$users[0]->id, $users[1]->id, $users[2]->id]));
-        // Utils::setAttributeAtXPath($process, '/bpmn:definitions/bpmn:process/bpmn:task[2]', 'pm:assignedUsers', implode(',', [$users[3]->id, $users[4]->id]));
-        // Utils::setAttributeAtXPath($process, '/bpmn:definitions/bpmn:process/bpmn:manualTask[1]', 'pm:assignedUsers', implode(',', [$users[5]->id, $users[6]->id]));
-        // Utils::setAttributeAtXPath($process, '/bpmn:definitions/bpmn:process/bpmn:manualTask[2]', 'pm:assignedUsers', implode(',', [$users[7]->id]));
-        // Utils::setAttributeAtXPath($process, '/bpmn:definitions/bpmn:process/bpmn:callActivity[1]', 'pm:assignedUsers', implode(',', [$users[8]->id, $users[9]->id]));
-        // Utils::setAttributeAtXPath($process, '/bpmn:definitions/bpmn:process/bpmn:callActivity[2]', 'pm:assignedUsers', implode(',', [$users[10]->id]));
-
-        // // Assign groups to process assignments
-        // Utils::setAttributeAtXPath($process, '/bpmn:definitions/bpmn:process/bpmn:task[1]', 'pm:assignedGroups', implode(',', [$groups[0]->id]));
-        // Utils::setAttributeAtXPath($process, '/bpmn:definitions/bpmn:process/bpmn:task[2]', 'pm:assignedGroups', implode(',', [$groups[1]->id, $groups[2]->id]));
-        // Utils::setAttributeAtXPath($process, '/bpmn:definitions/bpmn:process/bpmn:manualTask[1]', 'pm:assignedGroups', implode(',', [$groups[3]->id]));
-        // Utils::setAttributeAtXPath($process, '/bpmn:definitions/bpmn:process/bpmn:manualTask[2]', 'pm:assignedGroups', implode(',', [$groups[4]->id, $groups[5]->id]));
-        // Utils::setAttributeAtXPath($process, '/bpmn:definitions/bpmn:process/bpmn:callActivity[1]', 'pm:assignedGroups', implode(',', [$groups[6]->id, $groups[7]->id, $groups[8]->id]));
-        // Utils::setAttributeAtXPath($process, '/bpmn:definitions/bpmn:process/bpmn:callActivity[2]', 'pm:assignedGroups', implode(',', [$groups[9]->id]));
-    }
 }
